# openapi.json
{"openapi": "3.1.0", "info": {"title": "basuketto", "version": "0.1.0"}, "paths": {"/api/v1/utils/health-check/": {"get": {"tags": ["utils"], "summary": "Health Check", "operationId": "utils-health_check", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Response Utils-Health Check"}}}}}}}, "/api/v1/promos/": {"get": {"tags": ["promos"], "summary": "Read Promos", "description": "Retrieve promos.", "operationId": "promos-read_promos", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromosPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["promos"], "summary": "Create Promo", "description": "Create new promo.", "operationId": "promos-create_promo", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromoCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromoPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/promos/{id}": {"get": {"tags": ["promos"], "summary": "Read Promo", "description": "Get promo by ID.", "operationId": "promos-read_promo", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromoPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["promos"], "summary": "Update Promo", "description": "Update a promo.", "operationId": "promos-update_promo", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromoUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PromoPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["promos"], "summary": "Delete Promo", "description": "Delete an Promo.", "operationId": "promos-delete_promo", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"DiscountType": {"type": "string", "enum": ["percent", "fixed"], "title": "DiscountType"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Message": {"properties": {"message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["message"], "title": "Message"}, "PromoCreate": {"properties": {"code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Code"}, "discount_value": {"anyOf": [{"type": "number", "minimum": 0.0}, {"type": "string"}], "title": "Discount Value", "default": 0}, "discount_type": {"$ref": "#/components/schemas/DiscountType"}, "is_active": {"type": "boolean", "title": "Is Active"}}, "type": "object", "required": ["discount_type", "is_active"], "title": "PromoCreate"}, "PromoPublic": {"properties": {"code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Code"}, "discount_value": {"type": "string", "title": "Discount Value", "default": 0}, "discount_type": {"$ref": "#/components/schemas/DiscountType"}, "is_active": {"type": "boolean", "title": "Is Active"}, "id": {"type": "string", "format": "uuid", "title": "Id"}}, "type": "object", "required": ["discount_type", "is_active", "id"], "title": "PromoPublic"}, "PromoUpdate": {"properties": {"code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Code"}, "discount_value": {"anyOf": [{"type": "number", "minimum": 0.0}, {"type": "string"}, {"type": "null"}], "title": "Discount Value"}, "discount_type": {"anyOf": [{"$ref": "#/components/schemas/DiscountType"}, {"type": "null"}]}, "is_active": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Is Active"}}, "type": "object", "title": "PromoUpdate"}, "PromosPublic": {"properties": {"data": {"items": {"$ref": "#/components/schemas/PromoPublic"}, "type": "array", "title": "Data", "default": []}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["count"], "title": "PromosPublic"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}